+++ packages/frontend/src/features/auth/api/login.ts
@@ -1,29 +1,49 @@
-import type { LoginRequest } from "@pkg/shared";
-import { apiClient } from "../../../lib";
-import type { ApiResult } from "../../../lib/types";
-import type { LoginResult } from "./types/index";
+import type { LoginRequest } from "@pkg/shared"
+import { signIn } from "next-auth/react"
+import type { LoginResult } from "./types"
 
-export const login = async (body: LoginRequest): Promise<LoginResult> => {
-  const response: ApiResult = await apiClient({
-    path: "/api/auth/login",
-    method: "POST",
-    body: body
+
+export const login = async ({ email, password }: LoginRequest): Promise<LoginResult> => {
+  const result = await signIn("credentials", {
+    email,
+    password,
+    redirect: false, // SPAなのでリダイレクトは自前で制御
   });
 
-  if (!response.ok) {
-    console.error(response.error);
-    if (response.errorName === "UserUndefinedError") {
+
+  if (!result) {
+    return {
+      ok: false,
+      errorType: "Unknown"
+    }
+  }
+
+
+  if (!result.ok) {
+    if (result.status === 401) {
       return {
         ok: false,
         errorType: "UnRegistered"
       }
-    };
+    }
+
+    if (result.status === 403) {
+      console.error(`This error might be caused by CSRF.\n  errorMessage: ${result.error ?? "No error message found"}`);
+      return {
+        ok: false,
+        errorType: "Unknown"
+      }
+    }
+
     return {
       ok: false,
       errorType: "Unknown"
-    };
+    }
   }
 
-  console.info("Now logged in.");
-  return { ok: true };
+
+  console.info("You are logged in.");
+  return {
+    ok: true
+  }
 }
