+++ packages/frontend/src/features/auth/api/register.ts
@@ -1,11 +1,13 @@
 import { RegisterResponseSchema, type RegisterRequest } from "@pkg/shared";
-import { apiClient } from "../../../lib/api-client";
+import { login } from "./login";
 import type { ApiResult } from "../../../lib/types";
+import { apiClient } from "../../../lib";
 import type { RegisterResult } from "./types";
+
 
 export const register = async (body: RegisterRequest): Promise<RegisterResult> => {
   const response: ApiResult = await apiClient({
-    path: "/api/auth/register",
+    path: "/api/auth/v2/register",
     method: "POST",
     body: body
   });
@@ -30,6 +32,17 @@ export const register = async (body: RegisterRequest): Promise<RegisterResult> =
     return {
       ok: false,
       errorType: "GetTokenFailed"
+    }
+  }
+
+  const loginResult = await login({
+    email: body.email,
+    password: body.password
+  });
+  if (loginResult.ok === false) {
+    return {
+      ok: false,
+      errorType: "Unknown"
     }
   }
