--- packages/frontend/src/features/auth/api/register_v2.ts
@@ -1,50 +1,0 @@
-import { RegisterResponseSchema, type RegisterRequest } from "@pkg/shared";
-import type { RegisterResult } from "./types";
-import type { ApiResult } from "../../../lib/types";
-import { apiClient } from "../../../lib";
-import { login_v2 } from "./login_v2";
-
-
-export const register_v2 = async (body: RegisterRequest): Promise<RegisterResult> => {
-  const response: ApiResult = await apiClient({
-    path: "/api/auth/v2/register",
-    method: "POST",
-    body: body
-  });
-
-
-  if (!response.ok) {
-    console.error(response.error);
-    if (response.status === 409 || response.error.name === "AlreadyRegisteredError") {
-      return {
-        ok: false,
-        errorType: "AlreadyRegistered"
-      }
-    }
-
-    throw new Error("register failed with fetch Error");
-  }
-
-  const data = await response.body;
-  const parsedData = RegisterResponseSchema.safeParse(data);
-
-  if (!parsedData.success) {
-    return {
-      ok: false,
-      errorType: "GetTokenFailed"
-    }
-  }
-
-  const loginResult = await login_v2({
-    email: body.email,
-    password: body.password
-  });
-  if (loginResult.ok === false) {
-    return {
-      ok: false,
-      errorType: "Unknown"
-    }
-  }
-
-  return { ok: true }
-}
